SubmitButtonをcomponentにする
